Feature/s1 6.1 backend fixes updates

+++ BackEnd/src/main/java/com/unosquare/admin_core/back_end/dto/ClientDto.java
@@ -14,9 +14,9 @@ public class ClientDto implements java.io.Serializable {
     private String contactName;
     private String contactEmail;
     private String clientStatusDescription;
-    private short clientStatusId;
+    private int clientStatusId;
 
-    private short minimumEmployeesForTeam = 0;
+    private int minimumEmployeesForTeam = 0;
 
     public ClientDto() {
